+++ Source/CesiumRuntime/Private/Tests/CesiumFeatureIdTexture.spec.cpp
@@ -4,7 +4,6 @@
 #include "Misc/AutomationTest.h"
 
 using namespace CesiumGltf;
-using namespace CesiumGltf::MeshFeatures;
 
 BEGIN_DEFINE_SPEC(
     FCesiumFeatureIdTextureSpec,